+++ testing/api_test/Makefile
@@ -12,6 +12,6 @@ run_api_test:
 		--start_at=config_populator \
 		--task_config_uri=api_test_task_config.yaml \
 		--resource_config_uri=api_test_resource_config.yaml \
-		--extra_source_dir=testing/api_test/ \
+		--extra_source_dir=. \
 		--export_docker_artifact_registry=${DOCKER_ARTIFACT_REGISTRY} \
 		--wait
